Move GitHub Actions API client to gh-api directory
