+++ extensions/ql-vscode/src/remote-queries/export-results.ts
@@ -10,7 +10,7 @@ import {
 } from '../helpers';
 import { logger } from '../logging';
 import { QueryHistoryManager } from '../query-history';
-import { createGist } from './gh-actions-api-client';
+import { createGist } from './gh-api/gh-actions-api-client';
 import { RemoteQueriesManager } from './remote-queries-manager';
 import { generateMarkdown } from './remote-queries-markdown-generation';
 import { RemoteQuery } from './remote-query';
